Restrict image push to avoid errors on pull requests
According to [GitHub Action > Using encrypted secrets in a workflow](https://docs.github.com/en/actions/security-guides/encrypted-secrets#using-encrypted-secrets-in-a-workflow), secrets are not passed to the runner when a workflow is triggered from a forked repository. This means that pull requests will not be able to leverage the secrets to push images to the registry.

With this pull request, the push should be only performed in the following cases:

- the pull request is merged to the main or release branches,
- the workflow is manually triggered from the main or release branches,
- the scheduled time is reached.

Signed-off-by: Fabien Dupont <fabiendupont@pm.me>