Scan images with Anchore to the GH Action workflow
In this repository, we only build container image. The code is
maintained in a separate repository. So, we can only scan the generated
images for security flaws. This pull request adds an action that uses
the [Anchore
Scan](https://github.com/marketplace/actions/anchore-container-scan)
GitHub Action.

It also removes a remnant workflow.

Signed-off-by: Fabien Dupont <fabiendupont@pm.me>